fix(sum): fix Long-long problem both in test and src files

+++ week-04/day-03/test-exercises/src/PlayGround.java
@@ -9,10 +9,10 @@ public class PlayGround {
     return "apple";
   }
 
-  public Long sum(List<Integer> numbers) {
+  public long sum(List<Integer> numbers) {
     long result = 0;
     for (Integer num : numbers) {
-      result += num;
+      result += (long) num;
     }
     return result;
   }
