+++ week-04/day-03/test-exercises/tests/PlayGroundTest.java
@@ -1,6 +1,4 @@
 import org.junit.Test;
-import org.omg.PortableServer.LIFESPAN_POLICY_ID;
-
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
@@ -19,14 +17,14 @@ public class PlayGroundTest {
     PlayGround playGround = new PlayGround();
     List<Integer> numbers;
     numbers = Arrays.asList(4,5,4,3,2);
-    assertEquals((long) 18, (long) playGround.sum(numbers));
+    assertEquals(18, playGround.sum(numbers));
   }
 
   @Test
   public void sumWithNull() throws Exception {
     PlayGround playGround = new PlayGround();
     List<Integer> numbers = new ArrayList<>();
-    assertEquals((long) 0, (long) playGround.sum(numbers));
+    assertEquals(0, playGround.sum(numbers));
   }
 
   @Test
@@ -34,6 +32,6 @@ public class PlayGroundTest {
     PlayGround playGround = new PlayGround();
     List<Integer> numbers = new ArrayList<>();
     numbers.add(0);
-    assertEquals((long) 0, (long) playGround.sum(numbers));
+    assertEquals(0, playGround.sum(numbers));
   }
 }
